test(factory-sessions): keep SSE proof within package baseline
